refactor: 카운트 훅 분리
