+++ apps/landing/src/widgets/landing/ui/stats-section/stat-item.tsx
@@ -1,0 +1,17 @@
+import * as styles from './stat-item.css';
+
+interface StatItemProps {
+  value: string | number;
+  description: string;
+}
+
+const StatItem = ({ value, description }: StatItemProps) => (
+  <div className={styles.stat}>
+    <p className={styles.value}>
+      <span className={styles.accent}>{value}</span>%
+    </p>
+    <p className={styles.description}>{description}</p>
+  </div>
+);
+
+export default StatItem;
